rdf-to-ngsi-ld/pyproject.toml
[tool.poetry]
name = "rdf-to-ngsi-ld"
version = "0.1.0"
description = ""
authors = ["Ignacio Dominguez Martinez-Casanueva <ignacio.dominguezmartinez@telefonica.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
morph-kgc = "^2.6.3"
ngsi-ld-client = {git = "https://github.com/giros-dit/python-ngsi-ld-client.git", rev = "1.6.1"}
kafka-python = "^2.0.2"
pyoxigraph = "^0.3.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
